// frontend/src/api/model/githubCom4H1RZooraInternalDomainMediaOwner.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export interface GithubCom4H1RZooraInternalDomainMediaOwner {
  file_count?: number;
  name?: string;
  owner_id?: string;
  owner_kind?: string;
  total_size?: number;
}